+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- ClientRenderer: Use `<pre>` instead of `<div>` for diagram containers.
+
+### Added
+- ClientRenderer: Support changing the container tag
+  with the `ContainerTag` option.
+
 ## [0.3.0] - 2022-12-19
 ### Changed
 - Change the module path to `go.abhg.dev/goldmark/mermaid`.